requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
appdirs=1.4.3=pypi_0
ca-certificates=2020.1.1=0
certifi=2019.11.28=py36_0
dbus=1.13.12=h746ee38_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.63.1=h5a9c865_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=h9c2bf20_1
jpeg=9b=h024ee3a_2
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
ncurses=6.2=he6710b0_0
openssl=1.1.1d=h7b6447c_4
pcre=8.43=he6710b0_0
pip=20.0.2=py36_1
pycodestyle=2.5.0=pypi_0
pygubu=0.9.8.4=pypi_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.10=h0371630_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
setuptools=46.0.0=py36_0
sip=4.19.8=py36hf484d3e_0
sqlite=3.31.1=h7b6447c_0
tk=8.6.8=hbc83047_0
wheel=0.34.2=py36_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
